requirements.txt
Flask==2.2.2
gunicorn==20.1.0
line-bot-sdk==2.3.0
google-cloud-datastore==2.8.1
google-cloud-storage==2.5.0
Pillow==9.2.0
firebase-admin==5.4.0
pypinyin==0.47.1
requests==2.28.1
python-dotenv==0.21.0
